// frontend/src/generated/client/types.gen.ts
export type ProjectUpdate = {
    /**
     * Client Id
     */
    client_id?: string | null;
    /**
     * Name
     */
    name?: string | null;
    /**
     * Code
     */
    code?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Status
     */
    status?: 'active' | 'inactive' | null;
};